Add the ability to include additional proposal text
This adds a command-line option "-extract:proposals" which takes a list of
filenames and appends their text to the main ecma262/spec.html. This can
be used to run the analyzer on proposal texts.

(Although at this point it is probably very limited; it would need to
properly understand <ins> and <del> tags to work on many proposals.)