Parse Ecmarkup structured headers
The new-style structured headers used in Ecmarkup look like this:

```html
<emu-clause id="sec-esid" type="abstract operation">
  <h1>
    NameOfAbstractOperation (
      _parameter_: a type,
      optional _options_: another type, which may contain commas,
    )
  </h1>
  <dl class="header">
    <dt>description</dt>
    <dd>Description of the algorithm, to which Ecmarkup adds standard text
    before and after.</dd>
  </dl>
  <emu-alg>
    1. ...
  </emu-alg>
</emu-clause>
```

This adds initial support for parsing algorithms with structured headers.
The types are ignored, currently.

parseAoids() can select all AOs with structured headers by selecting on
the 'type' attribute in the <emu-clause>, which must be equal to "abstract
operation".

isBuiltin() can also use the 'type' attribute, for which we have to pass
the <emu-clause> element into the function. The previous method of
checking if the description contains "abstract operation" is no longer
sufficient because Ecmarkup now autogenerates that text when rendering.

Finally, this adds support for structured headers to the paramList parser.
Unfortunately the newlines are significant inside the <h1> element, so
this parser might not get all cases right. This uses a hacky definition of
structuredType to try to get it to eat all text until a comma followed by
an underscore (the beginning of a following parameter), a right bracket
(end of the parameter list), or the string "optional" (beginning of a
following optional parameter.)